Pulizia tabella MyCandle e MyBalance da record non collegati agli ordini

+++ Broker.Common/WebAPI/Binance/CancelOrder.cs
@@ -4,7 +4,7 @@ namespace Broker.Common.WebAPI.Binance
     internal class CancelOrder
     {
         public string symbol { get; set; }
-        public int orderId { get; set; }
+        public long orderId { get; set; }
         public string origClientOrderId { get; set; }
         public string clientOrderId { get; set; }
         public long transactTime { get; set; }
