+++ Broker.Common/WebAPI/Binance/PostOrder.cs
@@ -4,7 +4,7 @@ namespace Broker.Common.WebAPI.Binance
     internal class PostOrder
     {
         public string symbol { get; set; }
-        public int orderId { get; set; }
+        public long orderId { get; set; }
         public string clientOrderId { get; set; }
         public long transactTime { get; set; }
         public string price { get; set; }
